Restructure resources
This moves `levgrnd_bnds_generator.py`, which only serves as example
documentation, into a `docs/examples` directory, and converts it to an
Jupyter Notebook, which is a more appropriate format for this type of
documentation.

It also move the constant list `levgrnd_bnds` from the `levgrnd_bnds.py`
module, which was all the module contained, to a constant in the
`cmor_handlers` package, where it's used by one handler (but possibly
will be used by more).

This eliminates the need for the `resources` package entirely.